docs: clarify FAQ about does shorebird store my source code.
Shorebird stores more than just the dart binary, it stores the
whole .aab, .apk, .ipa, or .xcarchive to include assets and records
of other binary files so that it can later provide diffs to the
developer when patching a release with `shorebird patch`.

Also removed the double use of "store" meaning two different things
in one sentence and shortened the last sentence to avoid redundancy.