docs: tweek our linker status docs a little.
We don't have great data on how many apps are in the stores with
Shorebird. We do know that we have about 1000 monthly active
developers, and several thousand 3mo active developers, so guessing
there are around 1000 apps in the stores.  It's certainly in the
hundreds.

Also tweaked the wording and ordering a little.